Share more code between systest and MCP transports
Move window_properties() and take_snapshot_response() into
IntrospectionState so both transports call the same method instead of
duplicating the logic. Remove redundant index_to_handle/handle_to_index
wrappers and physical size/position helpers — import directly from
introspection. Move accessibility role mapping test to introspection.rs
where the function under test lives.
